+++ Aspid.Core.HSM/Assets/Plugins/Aspid/Core/HSM/Samples~/GameLoop/Controllers/GameHUDController.cs
@@ -1,0 +1,18 @@
+using Aspid.Core.HSM;
+using UnityEngine;
+
+namespace Aspid.Core.HSM.Samples.GameLoop.Controllers
+{
+    public class GameHUDController : IEnterController, IExitController
+    {
+        public void OnEnter()
+        {
+            Debug.Log("[HSM] GameHUD: HUD shown");
+        }
+
+        public void OnExit()
+        {
+            Debug.Log("[HSM] GameHUD: HUD hidden");
+        }
+    }
+}